+++ .prettierrc.js
@@ -1,2 +1,0 @@
-// @flow
-module.exports = require('@freighter/scripts/prettier-config');