+++ packages/eslint-config-freighter-repo/index.js
@@ -2,10 +2,6 @@
 module.exports = {
   overrides: [
     {
-      files: '**/jest.config.js',
-      env: { node: true },
-    },
-    {
       files: '**/__tests__/*.{js,ts}{x,}',
       env: { jest: true, node: true },
     },